# speed up forecast all