--- a/nowcasting_api/database_fast.py
+++ b/nowcasting_api/database_fast.py
@@ -15,6 +15,10 @@ def get_forecast_values_all_compact(
 ) -> [OneDatetimeManyForecastValues]:
     """ Get forecast values from the database.
 
+    We get all the latest forecast values for the blend model.
+    We convert the sqlalchemy objects to OneDatetimeManyForecastValues
+    Particular focus has been put on only get the data we need from the database. 
+
     This function
     1. get model ids
     2. get forecast values from the forecast_value_latest table.