--- a/nowcasting_api/gsp.py
+++ b/nowcasting_api/gsp.py
@@ -111,7 +111,8 @@ def get_all_available_forecasts(
             session=session,
             start_datetime_utc=start_datetime_utc,
             end_datetime_utc=end_datetime_utc,
-            gsp_ids=gsp_ids)
+            gsp_ids=gsp_ids,
+        )
 
     forecasts = get_forecasts_from_database(
         session=session,