# Pull Request

## Description

- speed up forecast all, this is just for compact=True and no creation time limit. This will help the UI.

Speed ups are
| route | current [s] | new [s] |
| --- | --- | --- |
| `/gsp/forecast/all/?compact=true`  | 3.67 | 1.15 |
| `/gsp/forecast/all/?compact=true&start_datetime_utc=2025-05-16T10:00:18&end_datetime_utc=2025-05-16T11:00:18`  | 10.7 | 1.17 |

Helps with #455 

## How Has This Been Tested?

- [x] Ci tests
- [x] ran locally

Ran UAT checks currently in https://github.com/openclimatefix/airflow-dags/pull/193, 
and two forecast/all checks go down from 9 seconds to ~1 second


## Checklist:

- [x] My code follows [OCF's coding style guidelines](https://github.com/openclimatefix/.github/blob/main/coding_style.md)
- [x] I have performed a self-review of my own code
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [x] I have checked my code and corrected any misspellings